+++ crates/fuel-core/src/service/sub_services.rs
@@ -40,10 +40,7 @@ use crate::{
 };
 use fuel_core_gas_price_service::v1::{
     algorithm::AlgorithmV1,
-    da_source_service::block_committer_costs::{
-        BlockCommitterDaBlockCosts,
-        BlockCommitterHttpApi,
-    },
+    da_source_service::block_committer_costs::BlockCommitterHttpApi,
     uninitialized_task::new_gas_price_service_v1,
 };
 use fuel_core_poa::{
@@ -189,8 +186,6 @@ pub fn init_sub_services(
     let metadata = StructuredStorage::new(database.gas_price().clone());
 
     let committer = BlockCommitterHttpApi::new(config.da_committer_url.clone());
-    let da_source = BlockCommitterDaBlockCosts::new(committer, None);
-    let on_chain_db = database.on_chain().clone();
 
     let gas_price_service = new_gas_price_service_v1(
         config.clone().into(),
@@ -198,10 +193,9 @@ pub fn init_sub_services(
         settings,
         block_stream,
         database.gas_price().clone(),
-        metadata,
         da_source,
         on_chain_db,
-        database.gas_price().clone(),
+        persisted_data,
     )?;
 
     let gas_price_provider = FuelGasPriceProvider::new(gas_price_service.shared.clone());