+++ crates/fuel-gas-price-algorithm/src/v1.rs
@@ -120,7 +120,7 @@ impl UnrecordedBlocks for BTreeMap<Height, Bytes> {
 /// instead of the actual profit. Setting the `avg_window` to 1 will effectively disable the
 /// moving average.
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq)]
-pub struct AlgorithmUpdaterV1<U> {
+pub struct AlgorithmUpdaterV1 {
     // Execution
     /// The gas price (scaled by the `gas_price_factor`) to cover the execution of the next block
     pub new_scaled_exec_price: u64,
@@ -164,8 +164,6 @@ pub struct AlgorithmUpdaterV1<U> {
     pub latest_da_cost_per_byte: u128,
     /// Activity of L2
     pub l2_activity: L2ActivityTracker,
-    /// The unrecorded blocks that are used to calculate the projected cost of recording blocks
-    pub unrecorded_blocks: U,
     /// Total unrecorded block bytes
     pub unrecorded_blocks_bytes: u128,
 }
@@ -290,10 +288,28 @@ impl L2ActivityTracker {
     pub fn current_activity(&self) -> u16 {
         self.chain_activity
     }
+
+    pub fn max_activity(&self) -> u16 {
+        self.max_activity
+    }
+
+    pub fn capped_activity_threshold(&self) -> u16 {
+        self.capped_activity_threshold
+    }
+
+    pub fn decrease_activity_threshold(&self) -> u16 {
+        self.decrease_activity_threshold
+    }
+
+    pub fn block_activity_threshold(&self) -> ClampedPercentage {
+        self.block_activity_threshold
+    }
 }
 
 /// A value that represents a value between 0 and 100. Higher values are clamped to 100
-#[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, PartialOrd)]
+#[derive(
+    serde::Serialize, serde::Deserialize, Debug, Copy, Clone, PartialEq, PartialOrd,
+)]
 pub struct ClampedPercentage {
     value: u8,
 }
@@ -320,28 +336,35 @@ impl core::ops::Deref for ClampedPercentage {
     }
 }
 
-impl<U: UnrecordedBlocks> AlgorithmUpdaterV1<U> {
-    pub fn update_da_record_data(
+impl AlgorithmUpdaterV1 {
+    pub fn update_da_record_data<U: UnrecordedBlocks>(
         &mut self,
         heights: &[u32],
         recorded_bytes: u32,
         recording_cost: u128,
+        unrecorded_blocks: &mut U,
     ) -> Result<(), Error> {
         if !heights.is_empty() {
-            self.da_block_update(heights, recorded_bytes as u128, recording_cost)?;
+            self.da_block_update(
+                heights,
+                recorded_bytes as u128,
+                recording_cost,
+                unrecorded_blocks,
+            )?;
             self.recalculate_projected_cost();
             self.update_da_gas_price();
         }
         Ok(())
     }
 
-    pub fn update_l2_block_data(
+    pub fn update_l2_block_data<U: UnrecordedBlocks>(
         &mut self,
         height: u32,
         used: u64,
         capacity: NonZeroU64,
         block_bytes: u64,
         fee_wei: u128,
+        unrecorded_blocks: &mut U,
     ) -> Result<(), Error> {
         let expected = self.l2_block_height.saturating_add(1);
         if height != expected {
@@ -372,7 +395,7 @@ impl<U: UnrecordedBlocks> AlgorithmUpdaterV1<U> {
             self.update_da_gas_price();
 
             // metadata
-            self.unrecorded_blocks.insert(height, block_bytes)?;
+            unrecorded_blocks.insert(height, block_bytes)?;
             self.unrecorded_blocks_bytes = self
                 .unrecorded_blocks_bytes
                 .saturating_add(block_bytes as u128);
@@ -533,13 +556,14 @@ impl<U: UnrecordedBlocks> AlgorithmUpdaterV1<U> {
             .saturating_div(100)
     }
 
-    fn da_block_update(
+    fn da_block_update<U: UnrecordedBlocks>(
         &mut self,
         heights: &[u32],
         recorded_bytes: u128,
         recording_cost: u128,
+        unrecorded_blocks: &mut U,
     ) -> Result<(), Error> {
-        self.update_unrecorded_block_bytes(heights);
+        self.update_unrecorded_block_bytes(heights, unrecorded_blocks);
 
         let new_da_block_cost = self
             .latest_known_total_da_cost_excess
@@ -561,10 +585,14 @@ impl<U: UnrecordedBlocks> AlgorithmUpdaterV1<U> {
 
     // Get the bytes for all specified heights, or get none of them.
     // Always remove the blocks from the unrecorded blocks so they don't build up indefinitely
-    fn update_unrecorded_block_bytes(&mut self, heights: &[u32]) {
+    fn update_unrecorded_block_bytes<U: UnrecordedBlocks>(
+        &mut self,
+        heights: &[u32],
+        unrecorded_blocks: &mut U,
+    ) {
         let mut total: u128 = 0;
         for expected_height in heights {
-            let res = self.unrecorded_blocks.remove(expected_height);
+            let res = unrecorded_blocks.remove(expected_height);
             match res {
                 Ok(Some(bytes)) => {
                     total = total.saturating_add(bytes as u128);